requirements.txt
torch
torchvision
timm
streamlit
pillow
pandas
matplotlib
numpypillow-heif
